+++ resources/[dg]/dg-vehicles/typescript/src/types/mechanic.d.ts
@@ -47,6 +47,7 @@ declare namespace Mechanic {
     items: {
       itemId: string;
       amount: number;
+      type: PartType;
     }[];
   };
 }