+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Keycard toegevoegd aan crafting table
 - UI Interactive app check werkt nu wel
 - Fix dat stresseffect actief blijft
+- Mechanic kan ticketjes voor tunes nu wel inbrengen
 
 ## [1.0.9]